+++ iso8601/date.go
@@ -338,3 +338,277 @@ func weeksInYear(year int) int {
 func isLeapYear(y int) bool {
 	return y%4 == 0 && (y%100 != 0 || y%400 == 0)
 }
+
+// DateLike defines an interface for date-related structures.
+// It provides methods for retrieving the date, validating the date,
+// and checking if the date is valid.
+type DateLike interface {
+	// Date returns the underlying Date value.
+	Date() Date
+
+	// IsValid checks whether the date is valid.
+	IsValid() bool
+
+	// Validate checks the correctness of the date and returns an error if it's invalid.
+	Validate() error
+}
+
+// Date represents a calendar date with year, month, and day components.
+type Date struct {
+	Year  int
+	Month time.Month
+	Day   int
+}
+
+var _ DateLike = Date{}
+
+// Date returns itself as it directly represents a date.
+func (d Date) Date() Date {
+	return d
+}
+
+// IsValid checks if the date is valid based on its year, month, and day values.
+func (d Date) IsValid() bool {
+	return d.Validate() == nil
+}
+
+// Validate checks the individual components of the date (year, month, and day)
+// and returns an error if any of them are out of the expected ranges.
+func (d Date) Validate() error {
+	if d.Year < 0 || d.Year > 9999 {
+		return &DateLikeRangeError{
+			Element: "year",
+			Value:   d.Year,
+			Year:    d.Year,
+			Min:     0,
+			Max:     9999,
+		}
+	}
+	if d.Month < 1 || d.Month > 12 {
+		return &DateLikeRangeError{
+			Element: "month",
+			Value:   int(d.Month),
+			Year:    d.Year,
+			Min:     1,
+			Max:     12,
+		}
+	}
+	daysInMonth := daysInMonth(d.Year, int(d.Month))
+	if d.Day < 1 || d.Day > daysInMonth {
+		return &DateLikeRangeError{
+			Element: "day of month",
+			Value:   d.Day,
+			Year:    d.Year,
+			Min:     1,
+			Max:     daysInMonth,
+		}
+	}
+	return nil
+}
+
+// StdTime converts the Date structure to a time.Time object, using UTC for the time.
+func (d Date) StdTime() time.Time {
+	return time.Date(d.Year, d.Month, d.Day, 0, 0, 0, 0, time.UTC)
+}
+
+// QuarterDate represents a date within a specific quarter of a year.
+// It includes the year, quarter (from 1 to 4), and day within that quarter.
+type QuarterDate struct {
+	Year    int
+	Quarter int
+	Day     int
+}
+
+var _ DateLike = QuarterDate{}
+
+// Date converts a QuarterDate into the standard Date representation.
+// It calculates the exact calendar date based on the year, quarter, and day within that quarter.
+func (q QuarterDate) Date() Date {
+	yday := q.Day // 1 ~ 366
+	for i := 1; i < q.Quarter; i++ {
+		yday += daysInQuarter(q.Year, i)
+	}
+	t := time.Date(q.Year, time.January, 1, 0, 0, 0, 0, time.UTC)
+	t = t.AddDate(0, 0, yday-1)
+	return Date{
+		Year:  t.Year(),
+		Month: t.Month(),
+		Day:   t.Day(),
+	}
+}
+
+// IsValid checks if the quarter date is valid based on its year, quarter, and day within the quarter values.
+func (q QuarterDate) IsValid() bool {
+	return q.Validate() == nil
+}
+
+// Validate checks the individual components of the quarter date (year, quarter, and day within the quarter)
+// and returns an error if any of them are out of the expected ranges.
+func (q QuarterDate) Validate() error {
+	if q.Year < 0 || q.Year > 9999 {
+		return &DateLikeRangeError{
+			Element: "year",
+			Value:   q.Year,
+			Year:    q.Year,
+			Min:     0,
+			Max:     9999,
+		}
+	}
+	if q.Quarter < 1 || q.Quarter > 4 {
+		return &DateLikeRangeError{
+			Element: "quarter",
+			Value:   q.Quarter,
+			Year:    q.Year,
+			Min:     1,
+			Max:     4,
+		}
+	}
+	daysInQuarter := daysInQuarter(q.Year, q.Quarter)
+	if q.Day < 1 || q.Day > daysInQuarter {
+		return &DateLikeRangeError{
+			Element: "day of quarter",
+			Value:   q.Day,
+			Year:    q.Year,
+			Min:     1,
+			Max:     daysInQuarter,
+		}
+	}
+	return nil
+}
+
+// WeekDate represents a date within a specific week of a given year,
+// following the ISO 8601 week-date system. It includes the year,
+// week number (1 to 52 or 53), and day of the week (1 for Monday to 7 for Sunday).
+type WeekDate struct {
+	Year int
+	Week int
+	Day  int
+}
+
+var _ DateLike = WeekDate{}
+
+// Date converts a WeekDate into the standard Date representation.
+// It calculates the exact calendar date based on the year, week number, and day of the week.
+func (w WeekDate) Date() Date {
+	// Find the first Thursday of the given year. This will be in the first week of the year according to ISO 8601.
+	thursday := time.Date(w.Year, time.January, 1, 0, 0, 0, 0, time.UTC)
+	for thursday.Weekday() != time.Thursday {
+		thursday = thursday.AddDate(0, 0, 1)
+	}
+
+	// Calculate the date of the Monday of week 1
+	monday := thursday.AddDate(0, 0, -3)
+
+	// Calculate the date corresponding to the given week and day
+	t := monday.AddDate(0, 0, (w.Week-1)*7+w.Day-1)
+	return Date{
+		Year:  t.Year(),
+		Month: t.Month(),
+		Day:   t.Day(),
+	}
+}
+
+// IsValid checks if the week date is valid based on its year, week number, and day of the week values.
+func (w WeekDate) IsValid() bool {
+	return w.Validate() == nil
+}
+
+// Validate checks the individual components of the week date (year, week number, and day of the week)
+// and returns an error if any of them are out of the expected ranges.
+func (w WeekDate) Validate() error {
+	if w.Year < 0 || w.Year > 9999 {
+		return &DateLikeRangeError{
+			Element: "year",
+			Value:   w.Year,
+			Year:    w.Year,
+			Min:     0,
+			Max:     9999,
+		}
+	}
+	if w.Day < 1 || w.Day > 7 {
+		return &DateLikeRangeError{
+			Element: "day of week",
+			Value:   int(w.Day),
+			Year:    w.Year,
+			Min:     1,
+			Max:     7,
+		}
+	}
+	weeksInYear := weeksInYear(w.Year)
+	if w.Week < 1 || w.Week > weeksInYear {
+		return &DateLikeRangeError{
+			Element: "week",
+			Value:   w.Week,
+			Year:    w.Year,
+			Min:     1,
+			Max:     weeksInYear,
+		}
+	}
+	return nil
+}
+
+// OrdinalDate represents a date specified by its year and the day-of-year (ordinal date),
+// where the day-of-year ranges from 1 through 365 (or 366 in a leap year).
+type OrdinalDate struct {
+	Year int
+	Day  int
+}
+
+var _ DateLike = OrdinalDate{}
+
+// Date converts an OrdinalDate into the standard Date representation.
+// It calculates the exact calendar date based on the year and the day-of-year.
+func (o OrdinalDate) Date() Date {
+	yday := o.Day // 1 ~ 366
+	t := time.Date(o.Year, time.January, 1, 0, 0, 0, 0, time.UTC)
+	t = t.AddDate(0, 0, yday-1)
+	return Date{
+		Year:  o.Year,
+		Month: t.Month(),
+		Day:   t.Day(),
+	}
+}
+
+// IsValid checks if the ordinal date is valid based on its year and day-of-year values.
+func (o OrdinalDate) IsValid() bool {
+	return o.Validate() == nil
+}
+
+// Validate checks the individual components of the ordinal date (year and day-of-year)
+// and returns an error if any of them are out of the expected ranges.
+func (o OrdinalDate) Validate() error {
+	if o.Year < 0 || o.Year > 9999 {
+		return &DateLikeRangeError{
+			Element: "year",
+			Value:   o.Year,
+			Year:    o.Year,
+			Min:     0,
+			Max:     9999,
+		}
+	}
+	daysInYear := daysInYear(o.Year)
+	if o.Day < 1 || o.Day > daysInYear {
+		return &DateLikeRangeError{
+			Element: "day of year",
+			Value:   o.Day,
+			Year:    o.Year,
+			Min:     1,
+			Max:     daysInYear,
+		}
+	}
+	return nil
+}
+
+// DateLikeRangeError indicates that a value is not in an expected range for DateLike.
+type DateLikeRangeError struct {
+	Element string
+	Value   int
+	Year    int
+	Min     int
+	Max     int
+}
+
+// Error implements the error interface.
+func (e *DateLikeRangeError) Error() string {
+	return fmt.Sprintf("iso8601: %d %s is not in range %d-%d in %d", e.Value, e.Element, e.Min, e.Max, e.Year)
+}
