+++ iso8601/time_test.go
@@ -559,3 +559,30 @@ func Test_parseExtendedTime(t *testing.T) {
 		})
 	}
 }
+
+func TestTimeRangeError_Error(t *testing.T) {
+	tests := []struct {
+		name string
+		err  *TimeRangeError
+		want string
+	}{
+		{
+			name: "valid error",
+			err: &TimeRangeError{
+				Element: "hour",
+				Value:   25,
+				Min:     0,
+				Max:     24,
+			},
+			want: "iso8601 time: 25 hour is not in range 0-24",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := tt.err.Error(); got != tt.want {
+				t.Errorf("TimeRangeError.Error() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
